# backend/crates/gateways/src/openapi.yaml
openapi: 3.1.0
info:
  title: Klynt Education Platform API
  version: 1.0.0
  description: Gateway API for the Klynt education platform.

servers:
  - url: http://localhost:3000/api/v1
    description: Local development gateway

paths:
  /health:
    get:
      summary: Health check
      tags:
        - Health
      responses:
        '200':
          description: Gateway is alive

  /auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Invalid input
        '409':
          description: Email already registered

  /auth/verify-email:
    post:
      summary: Verify email address
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Email verified successfully
        '400':
          description: Invalid or expired token

  /auth/request-password-reset:
    post:
      summary: Request password reset
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestPasswordResetRequest'
      responses:
        '200':
          description: Password reset email sent (if email exists)

  /auth/reset-password:
    post:
      summary: Reset password
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successfully
        '400':
          description: Invalid or expired token

  /auth/login:
    post:
      summary: Create session (login)
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid credentials

  /auth/logout:
    post:
      summary: End session (logout)
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: Logged out successfully

components:
  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        full_name:
          type: [string, "null"]

    VerifyEmailRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string

    RequestPasswordResetRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email

    ResetPasswordRequest:
      type: object
      required:
        - token
        - new_password
      properties:
        token:
          type: string
        new_password:
          type: string
          minLength: 8

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        remember_me:
          type: [boolean, "null"]

    LoginResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - expires_at
        - user
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_at:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/UserSessionInfo'

    UserSessionInfo:
      type: object
      required:
        - id
        - email
        - role
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: [string, "null"]
        role:
          type: string
          enum: [student, teacher, admin, parent]

    LogoutRequest:
      type: object
      required:
        - session_token
      properties:
        session_token:
          type: string

tags:
  - name: Health
    description: Operational health checks
  - name: Authentication
    description: User authentication and authorization